organized the files, added new imges and error handeling for signup
