+++ .expo/packager-info.json
@@ -1,9 +1,9 @@
 {
   "devToolsPort": 19002,
-  "expoServerPort": 19000,
-  "packagerPort": 19001,
-  "packagerPid": 25278,
+  "expoServerPort": null,
+  "packagerPort": null,
+  "packagerPid": null,
   "expoServerNgrokUrl": "https://uz-phz.anonymous.whitebrandchat.exp.direct",
   "packagerNgrokUrl": "https://packager.uz-phz.anonymous.whitebrandchat.exp.direct",
-  "ngrokPid": 25293
+  "ngrokPid": 37872
 }
