+++ package.json
@@ -16,6 +16,7 @@
     "react-dom": "^16.8.6",
     "react-native": "latest",
     "react-native-contacts": "^5.0.1",
+    "react-native-elements": "^1.1.0",
     "react-native-gifted-chat": "^0.9.11",
     "react-native-phone-input": "^0.2.2",
     "react-native-web": "^0.11.4",
