--- redux/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import currentUser from './currentUser';
-import login from './login';
-import myConversations from './myConversations';
-
-export default combineReducers({
-	currentUser,
-	login,
-	myConversations,
-});
