+++ src/config/ImageAssets.js
@@ -1,0 +1,11 @@
+const ImageAssets = {
+	WHITE_BRAND_ICON: require('../assets/chat.png'),
+	MOBILE_NUMBER_ICON: require('../assets/mobile-number.png'),
+	ERROR_404_ICON: require('../assets/network-error.png'),
+	PINCODE_ICON: require('../assets/pincode.png'),
+	EMPTY_INBOX_ICON: require('../assets/no-messages.png'),
+	PLUS_ICON: require('../assets/plus.png'),
+	TEXT_MESSAGE_ICON: require('../assets/text.png'),
+};
+
+export default ImageAssets;
