+++ src/config/index.js
@@ -1,0 +1,5 @@
+import ImageAssets from './ImageAssets';
+
+export default {
+	ImageAssets,
+};
