+++ src/helpers/AppNavigation.js
@@ -1,9 +1,8 @@
 import { createStackNavigator, createAppContainer } from 'react-navigation';
-import HomeScreen from '../screens/HomeScreen';
-import LoginScreen from '../screens/LoginScreen';
-import SplashScreen from '../screens/SplashScreen';
-import ContactsScreenStyles from '../styling/ContactsScreen.styles';
-import ContactsScreen from '../screens/ContactsScreen';
+import { HomeScreen } from '../modules/Home';
+import { LoginScreen } from '../modules/Login';
+import { SplashScreen } from '../modules/Splash';
+import { ContactsScreen } from '../modules/Contacts';
 
 const AppNavigator = createStackNavigator(
 	{
