+++ src/modules/Contacts/index.js
@@ -1,0 +1,5 @@
+import ContactsScreen from './components/ContactsScreen';
+import actions from './actions';
+import reducer from './reducer';
+
+export { actions, ContactsScreen, reducer };
