+++ src/modules/Home/index.js
@@ -1,0 +1,5 @@
+import HomeScreen from './components/HomeScreen';
+import actions from './actions';
+import reducer from './reducer';
+
+export { HomeScreen, reducer, actions };
