+++ src/modules/Login/index.js
@@ -1,0 +1,5 @@
+import LoginScreen from './components/LoginScreen';
+import actions from './actions';
+import reducer from './reducer';
+
+export { LoginScreen, reducer, actions };
