+++ src/modules/Login/reducer.js
@@ -1,20 +1,23 @@
 import {
 	SET_PHONE_NUMBER,
 	SET_SECURITY_CODE,
-	JWT_REHIDRATION_SUCCESSFUL,
-	JWT_REHIDRATION_FAILED,
-} from '../actions/login';
-import {
+	USER_LOGIN_SUCCESS,
 	SECURITY_CODE_SENT,
 	TOGGLE_LOGIN_LOADING_STATE,
-	USER_LOGIN_SUCCESS,
-} from '../actions/user';
+	USER_SIGNUP_FAILED,
+} from './actions';
+
+import {
+	JWT_REHIDRATION_SUCCESSFUL,
+	JWT_REHIDRATION_FAILED,
+} from '../Auth/actions';
 
 let initialState = {
 	phoneNumber: '',
 	securityCode: '',
 	codeSent: null,
 	loading: false,
+	signupError: null,
 	rehidratingJwt: true,
 };
 
@@ -28,8 +31,10 @@ export default function(state = initialState, { type, payload }) {
 			return { ...state, phoneNumber: payload };
 		case SET_SECURITY_CODE:
 			return { ...state, securityCode: payload };
+		case USER_SIGNUP_FAILED:
+			return { ...state, signupError: true, loading: false };
 		case SECURITY_CODE_SENT:
-			return { ...state, codeSent: true };
+			return { ...state, codeSent: true, signupError: null };
 		case TOGGLE_LOGIN_LOADING_STATE:
 			return { ...state, loading: !state.loading };
 		case USER_LOGIN_SUCCESS:
