+++ src/modules/Splash/index.js
@@ -1,0 +1,3 @@
+import SplashScreen from './components/SplashScreen';
+
+export { SplashScreen };
