+++ src/redux/index.js
@@ -1,0 +1,4 @@
+import store from './store';
+import rootReducer from './rootReducer';
+
+export { store, rootReducer };
