+++ src/redux/store.js
@@ -1,5 +1,5 @@
 import { createStore, compose, applyMiddleware } from 'redux';
-import reducers from './reducers';
+import reducers from './rootReducer';
 import ReduxThunk from 'redux-thunk';
 
 const devTools = window.devToolsExtension ? window.devToolsExtension() : f => f;
